+++ spec/models/activity_spec.rb
@@ -480,6 +480,31 @@ RSpec.describe Activity, type: :model do
       end
     end
 
+    context "when activity is not a fund" do
+      context "and is 'ingested: false'" do
+        context "and the collaboration_type is blank" do
+          subject(:activity) { build(:programme_activity, collaboration_type: nil, ingested: false) }
+          it "should not be valid" do
+            expect(activity.valid?(:collaboration_type_step)).to be_falsey
+          end
+        end
+      end
+
+      context "but the activity is 'ingested:true'" do
+        subject(:activity) { build(:programme_activity, collaboration_type: nil, ingested: true) }
+        it "should be valid" do
+          expect(activity.valid?(:collaboration_type_step)).to be_truthy
+        end
+      end
+    end
+
+    context "when activity is a fund and collaboration_type is blank" do
+      subject(:activity) { build(:fund_activity, collaboration_type: nil) }
+      it "should be valid" do
+        expect(activity.valid?(:collaboration_type_step)).to be_truthy
+      end
+    end
+
     context "when flow is blank" do
       subject(:activity) { build(:activity, flow: nil) }
       it "should not be valid" do