+++ CHANGELOG.md
@@ -324,6 +324,7 @@
 - Update the budget policy
 - Report variance is shown in a tab
 - Show budgets added in a report on the reports view
+- `Collaboration type` form field added to create activity journey and activity XML
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-16...HEAD
 [release-16]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-15...release-16